Setup reducer, photo reducer, reselect
