+++ web/package.json
@@ -35,6 +35,7 @@
         "@types/reach__router": "^1.2.4",
         "@types/react": "^16.8.22",
         "@types/react-dom": "^16.8.4",
+        "@types/reselect": "^2.2.0",
         "antd": "^3.19.8",
         "axios": "^0.19.0",
         "blockstack": "^19.2.1",
@@ -62,6 +63,9 @@
         "react-dom": "^16.8.6",
         "react-helmet": "^5.2.0",
         "react-image-lightbox": "^5.1.0",
+        "react-redux": "^7.1.1",
+        "redux": "^4.0.4",
+        "reselect": "^4.0.0",
         "rxjs": "^6.5.2",
         "uuid": "^3.3.2"
     },
@@ -73,6 +77,7 @@
         "@babel/core": "^7.4.5",
         "@storybook/react": "^5.1.9",
         "@types/file-saver": "^2.0.1",
+        "@types/react-redux": "^7.1.2",
         "@typescript-eslint/eslint-plugin": "^2.0.0",
         "@typescript-eslint/parser": "^2.0.0",
         "babel-loader": "^8.0.6",
@@ -91,7 +96,7 @@
         "npm-run-all": "^4.1.5",
         "prettier": "^1.18.2",
         "prettier-eslint": "^9.0.0",
-        "typescript": "^3.5.2",
+        "typescript": "^3.6.2",
         "workerize-loader": "^1.1.0"
     },
     "repository": {
