+++ web/src/app/redux/root.reducer.ts
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import photoReducer from './photo/photo.reducer';
+
+export default combineReducers({
+    photo: photoReducer,
+});
