+++ web/src/app/redux/root.store.ts
@@ -1,0 +1,4 @@
+import { createStore } from 'redux';
+import rootReducer from './root.reducer'
+
+export default createStore(rootReducer)
